PN-16770

+++ PAS_Database/dbo/Tables/Tables1/ReceivingReconciliationDetails.sql
@@ -29,7 +29,7 @@
     [GlAccountName]                   VARCHAR (200)   NULL,
     [Type]                            INT             NULL,
     [StockType]                       VARCHAR (50)    NULL,
-    [RemainingRRQty]                  INT             NULL,
+    [RemainingRRQty]                  DECIMAL (18, 6) NULL,
     [FreightAdjustment]               DECIMAL (18, 2) NULL,
     [TaxAdjustment]                   DECIMAL (18, 2) NULL,
     [FreightAdjustmentPerUnit]        DECIMAL (18, 2) NULL,
@@ -40,6 +40,8 @@
     [VendorProformaInvoiceId]         BIGINT          NULL,
     CONSTRAINT [PK_ReceivingReconciliationDetails] PRIMARY KEY CLUSTERED ([ReceivingReconciliationDetailId] ASC)
 );
+
+
